Move catalog endpoints to a separate file
